feat: 웹툰, 웹소설 관심작품 등록 구현

+++ pageon_backend/src/main/java/com/pageon/backend/entity/Like.java
@@ -28,7 +28,7 @@ public class Like extends BaseTimeEntity {
 
     @Enumerated(EnumType.STRING)
     private ContentType contentType;     // 좋아요 작품이 웹툰인지 웹소설인지 표시
-    private Integer contentId;      // 해당 작품의 id
+    private Long contentId;      // 해당 작품의 id
 
 
 }
